+++ flake.nix
@@ -20,6 +20,8 @@
                     cmake
                     pkg-config
 
+                    git
+                    unzip
                     p7zip
                     libwebp
                 ];